+++ cognite/powerops/resync/models/_shared_v1_v2/cogshop_model.py
@@ -13,6 +13,7 @@ from typing_extensions import TypeAlias
 from cognite.powerops.client._generated.cogshop1 import data_classes as cogshop_v1
 from cognite.powerops.prerun_transformations.transformations import Transformation as TransformationV2
 from cognite.powerops.resync import config
+from cognite.powerops.resync.config import Transformation
 from cognite.powerops.resync.models.base import CDFFile, Model
 from cognite.powerops.utils.serialization import load_yaml
 
@@ -108,7 +109,7 @@ def _to_model_without_timeseries(
 
     model_dict = {"model": model}
 
-    # TODO: remove this when standardizing input SHOP file structure (model vs model + connections)
+    # TODO: remove this when standardizing parameters SHOP file structure (model vs model + connections)
     if "connections" in model_incl_time_and_connections:
         model_dict["connections"] = model_incl_time_and_connections["connections"]
 
@@ -125,16 +126,119 @@ def _create_transformation(order: int, transformation: dict | config.Transformat
     )
 
 
-def _create_transformationV2(order: int, transformation: dict | TransformationV2) -> cogshop_v1.TransformationApply:
+def _create_transformationV2(
+    order: int, transformation: dict | TransformationV2 | Transformation
+) -> cogshop_v1.TransformationApply:
     """
     Adapter betweeen transformationsV2 pydantic instances to CogShop model 1 FDM instances
     """
     if isinstance(transformation, dict):
         transformation = TransformationV2.load(transformation)
-    dumped_input_args = json.dumps(transformation.input_to_dict() or {}, separators=(",", ":"), ensure_ascii=False)
-    external_id = f"Tr_{transformation.name()}_{dumped_input_args}_{order}"
-    if len(external_id) > 255:
-        external_id = f"Tr_{transformation.name()}_{dumped_input_args[:len(dumped_input_args) // 2]}_{order}"
-    return cogshop_v1.TransformationApply(
-        external_id=external_id, method=transformation.name(), arguments=dumped_input_args, order=order
+    # TODO: temp hack to be able to read in old transformations from price scenarios in Dayahead and Rkom
+    # TODO: remove once market configs can load the new price scenarios config
+    elif isinstance(transformation, Transformation):
+        transformation = transformations_v2_transformer(transformation)
+
+    dumped_parameters_args = json.dumps(
+        transformation.parameters_to_dict() or {}, separators=(",", ":"), ensure_ascii=False
     )
+    external_id = f"Tr2_{transformation.name}_{dumped_parameters_args}_{order}"
+    if len(external_id) > 255:
+        external_id = f"Tr2_{transformation.name}_{dumped_parameters_args[:len(dumped_parameters_args) // 2]}_{order}"
+    return cogshop_v1.TransformationApply(
+        external_id=external_id, method=transformation.name, arguments=dumped_parameters_args, order=order
+    )
+
+
+def transformations_v2_transformer(
+    transformation: Transformation | dict, object_type: Optional[str] = None, object_name: Optional[str] = None
+) -> TransformationV2:
+    """
+    Adapter that converts old transformation to an instance of new transformationsV2.
+    From these instances, the transformationsV2 to FDM adapter "_create_transformationV2"
+    can be called to create DM nodes and edges
+    """
+    if isinstance(transformation, dict):
+        transformation = config.Transformation(**transformation)
+
+    transformation_dict = {}
+    if transformation.transformation.name == "ADD":
+        transformation_dict = {"AddConstant": {"parameters": {"constant": transformation.kwargs.get("value")}}}
+    elif transformation.transformation.name in ["ADD_FROM_OFFSET", "DYNAMIC_ADD_FROM_OFFSET"]:
+        transformation_dict = {
+            "AddFromOffset": {
+                "parameters": {
+                    "relative_datapoints": [
+                        {"offset_minute": k, "offset_value": v} for k, v in transformation.kwargs.items()
+                    ]
+                }
+            }
+        }
+    elif transformation.transformation.name == "ADD_WATER_IN_TRANSIT":
+        gate_or_plant = "gate" if "gate_name" in transformation.kwargs.keys() else "plant"
+        transformation_dict = {
+            "AddWaterInTransit": {
+                "parameters": {
+                    "discharge_ts_external_id": transformation.kwargs.get("external_id"),
+                    "transit_object_type": gate_or_plant,
+                    "transit_object_name": transformation.kwargs.get(f"{gate_or_plant}_name"),
+                }
+            }
+        }
+    elif transformation.transformation.name == "ADD_WATER_IN_TRANSIT":
+        gate_or_plant = "gate" if "gate_name" in transformation.kwargs.keys() else "plant"
+        transformation_dict = {
+            "AddWaterInTransit": {
+                "parameters": {
+                    "discharge_ts_external_id": transformation.kwargs.get("external_id"),
+                    "transit_object_type": gate_or_plant,
+                    "transit_object_name": transformation.kwargs.get(f"{gate_or_plant}_name"),
+                }
+            }
+        }
+
+    elif transformation.transformation.name == "MULTIPLY":
+        transformation_dict = {"MultiplyConstant": {"parameters": {"constant": transformation.kwargs.get("value")}}}
+    elif transformation.transformation.name == "MULTIPLY_FROM_OFFSET":
+        transformation_dict = {
+            "MultiplyFromOffset": {
+                "parameters": {
+                    "relative_datapoints": [
+                        {"offset_minute": k, "offset_value": v} for k, v in transformation.kwargs.items()
+                    ]
+                }
+            }
+        }
+    elif transformation.transformation.name == "TO_BOOL":
+        transformation_dict = {"ToBool": None}
+    elif transformation.transformation.name == "TO_INT":
+        transformation_dict = {"ToInt": None}
+    elif transformation.transformation.name in [
+        "ZERO_IF_NOT_ONE",
+        "GATE_SCHEDULE_FLAG_VALUE_MAPPING",
+        "GENERATOR_PRODUCTION_SCHEDULE_FLAG_VALUE_MAPPING",
+        "PLANT_PRODUCTION_SCHEDULE_FLAG_VALUE_MAPPING",
+    ]:
+        transformation_dict = {"ZeroIfNotOne": None}
+    elif transformation.transformation.name == "ONE_IF_TWO":
+        transformation_dict = {"OneIfTwo": None}
+    elif transformation.transformation.name in ["STATIC", "DYNAMIC_STATIC"]:
+        transformation_dict = {
+            "StaticValues": {
+                "parameters": {
+                    "relative_datapoints": [
+                        {"offset_minute": k, "offset_value": v} for k, v in transformation.kwargs.items()
+                    ]
+                }
+            }
+        }
+    elif transformation.transformation.name == "RESERVOIR_LEVEL_TO_VOLUME":
+        transformation_dict = {
+            "HeightToVolume": {"parameters": {"object_type": object_type, "object_name": object_name}}
+        }
+    elif transformation.transformation.name in ["DO_NOTHING", "GATE_OPENING_METER_TO_PERCENT"]:
+        transformation_dict = {"DoNothing": None}
+    else:
+        raise NotImplementedError(f"Unknown transformation {transformation.transformation.name}")
+
+    return TransformationV2.load(transformation_dict)